llm/requirements.txt
groq
pydantic-settings

# server
fastapi
uvicorn[standard]

# communication
pydantic
httpx
wizard-contest-commonlib==0.0.9